+++ app/src/main/java/nextstep/jwp/web/model/User.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.model;
+package nextstep.jwp.web.model;
 
 public class User {
