+++ app/src/test/java/nextstep/jwp/http/RequestHandlerTest.java
@@ -1,14 +1,13 @@
-package nextstep.jwp;
+package nextstep.jwp.http;
 
-import nextstep.jwp.http.RequestHandler;
-import org.junit.jupiter.api.Test;
+import static org.assertj.core.api.Assertions.assertThat;
 
 import java.io.File;
 import java.io.IOException;
 import java.net.URL;
 import java.nio.file.Files;
-
-import static org.assertj.core.api.Assertions.assertThat;
+import nextstep.jwp.MockSocket;
+import org.junit.jupiter.api.Test;
 
 class RequestHandlerTest {
 
@@ -34,7 +33,7 @@ class RequestHandlerTest {
     @Test
     void index() throws IOException {
         // given
-        final String httpRequest= String.join("\r\n",
+        final String httpRequest = String.join("\r\n",
                 "GET /index.html HTTP/1.1 ",
                 "Host: localhost:8080 ",
                 "Connection: keep-alive ",
@@ -52,7 +51,7 @@ class RequestHandlerTest {
         String expected = "HTTP/1.1 200 OK \r\n" +
                 "Content-Type: text/html;charset=utf-8 \r\n" +
                 "Content-Length: 5564 \r\n" +
-                "\r\n"+
+                "\r\n" +
                 new String(Files.readAllBytes(new File(resource.getFile()).toPath()));
         assertThat(socket.output()).isEqualTo(expected);
     }
